+++ CHANGES.rst
@@ -20,6 +20,12 @@ Changelog
   - Adapted `MeetingItem.validate_category` and `MeetingItem.validate_classifier`
     to make sure passed value exists.
     [gbastien]
+  - Added `Migrator.addCKEditorStyle` helper to ease adding an new CKeditor style.
+    [gbastien]
+  - Added possibility to anonymize a part of a rich text using new added CKeditor
+    style `span.pm-anonymize`.
+    This is also taken into account when data get using restapi.
+    [gbastien]
 
 4.1.28.28 (2021-08-17)
 ----------------------